+++ src/aa_utilities/graphics/__init__.py
@@ -1,6 +1,6 @@
 
 from ._graphics import *
-from ._heatmap import heatmap
+from ._heatmap import heatmap, overlay_boxes
 from ._interactive_selection import InteractiveSelection
 from ._repel_text import repel_text
 from ._upset_plot import UpsetPlot