+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added:
 - `helper.pandas.reorder_by_similarity` to reorder rows/columns by row/column similarity
 - `.computation.modeling.stability_selection` is now updated to accept any type of output.
+- `.graphics.overlay_boxes` to overlay boxes over existing `sns.clustermap` objects
 
 ## [0.0.44] - 2026-04-13
 ### Added: